Fixed a bug with fillGridWithTriangeSurfaceVoxels
When working in a 2D plane I find the first voxel that collides with a line segment of the triangle, and then I walk to the back of the grid and attempt to "bound" the triangle from the other side. I was using the wrong y value when generating the grid.

The backwards iterating y-value is the one we should use to do triangleAABB collisions.